Add a remove function for unstaging files
Additonally, pretty much all the other code got a general clean-up. For
example, instead of each function checking if the program is running in
an AIT repo, this is done initially in ait.go. Status no longer returns
an error because it doesn't need to. A utility function was added to
root.go which returns the size of a file.