Tweak PathMatch to work almost perfectly
From a shell, PathMatch works just as you would expect it to. When wildcards aren't expanded by a shell however, like if you're passing the arguments to add/rm through an IDE,
The behavior is slightly different than if you were using a shell. There is a lengthy comment on AddRun that explains this in more detail.