Add submit.go with add, commit, and push functions
There is a LOT of work to do on these functions yet. However, they outline the basic idea which is that with one command from the user, ait does 3 underlying git commands. Also made a function that opens files and log.Fatals any error in utils.go because I noticed that was repeated a lot.