add rudimentary submit function
this function takes a URL to a git repo, clones it, generates a keyset
and puts it in the repo, commits it with a user-written message, and
pushes it with the user's Github credentals (I think we're only
supporting github). There is much more to be done. Subject line (try to keep under 50 characters)