Improve the submit function
Submit now has somewhat better error handling, in that if it encounters an error, the repo that was being worked on is deleted. It's a bit of a hamfisted approach, and I may look to improve it in the future
but users really aren't supposed to be interacting with git and/or their repos directly. Also, when entering their passwords, users will no longer see their input echoed back at them. Finally, I
renamed commit_messsage.go to commit_message.go