Add the first half of pull request support
As of now, repos can be forked and forks can be cloned to local storage. The submit process in general has been made more user friendly with explanations of things that may not be
readily apparent, like how to create an Oauth token. Also a new function was added to utils which extracts the owner of a repo from its url.